+++ equinox-projector/Consumer/Infrastructure.fs
@@ -1,11 +1,11 @@
 ﻿[<AutoOpen>]
-module private ProjectorTemplate.Consumer.Infrastructure
+module ProjectorTemplate.Consumer.Infrastructure
 
 open System
 open System.Threading
 open System.Threading.Tasks
 
-type Async with
+type FSharp.Control.Async with
     static member AwaitTaskCorrect (task : Task<'T>) : Async<'T> =
         Async.FromContinuations <| fun (k,ek,_) ->
             task.ContinueWith (fun (t:Task<'T>) ->
@@ -17,6 +17,17 @@ type Async with
                 elif t.IsCompleted then k t.Result
                 else ek(Exception "invalid Task state!"))
             |> ignore
+//    static member AwaitTaskCorrect (task : Task) : Async<unit> =
+//        Async.FromContinuations <| fun (k,ek,_) ->
+//            task.ContinueWith (fun (t:Task) ->
+//                if t.IsFaulted then
+//                    let e = t.Exception
+//                    if e.InnerExceptions.Count = 1 then ek e.InnerExceptions.[0]
+//                    else ek e
+//                elif t.IsCanceled then ek (TaskCanceledException("Task wrapped with Async has been cancelled."))
+//                elif t.IsCompleted then k ()
+//                else ek(Exception "invalid Task state!"))
+//            |> ignore
 
 type System.Threading.SemaphoreSlim with
     /// F# friendly semaphore await function
@@ -26,7 +37,7 @@ type System.Threading.SemaphoreSlim with
         let task = semaphore.WaitAsync(timeout, ct) 
         return! Async.AwaitTaskCorrect task
     }
-    /// Throttling wrapper which waits asynchronously until the semaphore has available capacity
+    /// Throttling wrapper that waits asynchronously until the semaphore has available capacity
     member semaphore.Throttle(workflow : Async<'T>) : Async<'T> = async {
         let! _ = semaphore.Await()
         try return! workflow