+++ equinox-sync/Sync/Program.fs
@@ -1,11 +1,9 @@
 ﻿module SyncTemplate.Program
 
 open Equinox.Cosmos
-#if cosmos
-open Equinox.Cosmos.Projection
-#else
 open Equinox.EventStore
-#endif
+open Propulsion.Cosmos
+open Propulsion.EventStore
 open Serilog
 open System
 
@@ -21,15 +19,16 @@ module CmdParser =
     [<NoEquality; NoComparison>]
     type Parameters =
         | [<MainCommand; ExactlyOnce>] ConsumerGroupName of string
-        | [<AltCommandLine "-S"; Unique>] LocalSeq
-        | [<AltCommandLine "-v"; Unique>] Verbose
         | [<AltCommandLine "-z"; Unique>] FromTail
         | [<AltCommandLine "-r"; Unique>] MaxPendingBatches of int
         | [<AltCommandLine "-w"; Unique>] MaxWriters of int
-        | [<AltCommandLine "-cc"; Unique>] MaxCosmosConnections of int
+        | [<AltCommandLine "-mc"; Unique>] MaxConnections of int
+        | [<AltCommandLine "-ms"; Unique>] MaxSubmit of int
+        | [<AltCommandLine "-S"; Unique>] LocalSeq
+        | [<AltCommandLine "-v"; Unique>] Verbose
 #if cosmos
-        | [<AltCommandLine "-mp"; Unique>] MaxProcessing of int
         | [<AltCommandLine "-md"; Unique>] MaxDocuments of int
+        | [<AltCommandLine "-l"; Unique>] LagFreqM of float
         | [<AltCommandLine "-vc"; Unique>] ChangeFeedVerbose
         | [<AltCommandLine "-as"; Unique>] LeaseCollectionSource of string
         | [<AltCommandLine "-ad"; Unique>] LeaseCollectionDestination of string
@@ -51,19 +50,16 @@ module CmdParser =
             member a.Usage =
                 match a with
                 | ConsumerGroupName _ ->    "Projector consumer group name."
-                | LocalSeq ->               "configures writing to a local Seq endpoint at http://localhost:5341, see https://getseq.net"
-                | Verbose ->                "request Verbose Logging. Default: off"
                 | MaxWriters _ ->           "maximum number of concurrent writes to target permitted. Default: 512"
+                | MaxConnections _ ->       "size of Sink connection pool to maintain. Default: 1"
+                | MaxPendingBatches _ ->    "maximum number of batches to let processing get ahead of completion. Default: 16"
+                | MaxSubmit _ ->            "maximum number of batches to submit concurrently. Default: 8"
                 | MaxCosmosConnections _ -> "size of CosmosDb connection pool to maintain. Default: 1"
 #if cosmos
-                | MaxPendingBatches _ ->    "maximum number of batches to let processing get ahead of completion. Default: 32"
-                | MaxProcessing _ ->        "maximum number of batches to submit concurrently. Default: 16"
                 | ChangeFeedVerbose ->      "request Verbose Logging from ChangeFeedProcessor. Default: off"
                 | FromTail _ ->             "(iff the Consumer Name is fresh) - force skip to present Position. Default: Never skip an event."
                 | MaxDocuments _ ->         "maximum item count to request from feed. Default: unlimited"
-                | LeaseCollectionSource _ ->"specify Collection Name for Leases collection, within `source` connection/database (default: `source`'s `collection` + `-aux`)."
-                | LeaseCollectionDestination _ -> "specify Collection Name for Leases collection, within [destination] `cosmos` connection/database (default: defined relative to `source`'s `collection`)."
-                | LagFreqS _ ->             "specify frequency to dump lag stats. Default: off"
+                | LagFreqM _ ->             "frequency (in minutes) to dump lag stats. Default: off"
                 | Source _ ->               "CosmosDb input parameters."
 #else
                 | MaxPendingBatches _ ->    "Maximum number of batches to let processing get ahead of completion. Default: 2048"
@@ -84,14 +80,14 @@ module CmdParser =
         member __.MaybeSeqEndpoint =    if a.Contains LocalSeq then Some "http://localhost:5341" else None
         member __.Verbose =             a.Contains Verbose
         member __.MaxWriters =          a.GetResult(MaxWriters,1024)
-        member __.CosmosConnectionPool =a.GetResult(MaxCosmosConnections,1)
+        member __.ConnectionPoolSize =  a.GetResult(MaxCosmosConnections,1)
 #if cosmos
-        member __.MaxPendingBatches =   a.GetResult(MaxPendingBatches,32)
+        member __.MaxReadAhead =        a.GetResult(MaxPendingBatches,32)
         member __.MaxProcessing =       a.GetResult(MaxProcessing,16)
         member __.ChangeFeedVerbose =   a.Contains ChangeFeedVerbose
         member __.LeaseId =             a.GetResult ConsumerGroupName
         member __.MaxDocuments =        a.TryGetResult MaxDocuments
-        member __.LagFrequency =        a.TryGetResult LagFreqS |> Option.map TimeSpan.FromSeconds
+        member __.LagFrequency =        a.TryGetResult LagFreqM |> Option.map TimeSpan.FromMinutes
 #else
         member __.MaxPendingBatches =   a.GetResult(MaxPendingBatches,2048)
         member __.VerboseConsole =      a.Contains VerboseConsole
@@ -107,7 +103,7 @@ module CmdParser =
 #endif
 
         member val Source : SourceArguments = SourceArguments(a.GetResult Source)
-        member __.Destination : DestinationArguments = __.Source.Destination
+        member __.Sink : Choice<CosmosSinkArguments,EsSinkArguments> = __.Source.Sink
 #if cosmos
         member x.BuildChangeFeedParams() =
             let disco, db =
@@ -158,7 +154,8 @@ module CmdParser =
 #endif
         | [<AltCommandLine "-e">] CategoryBlacklist of string
         | [<AltCommandLine "-i">] CategoryWhitelist of string
-        | [<CliPrefix(CliPrefix.None); Unique(*ExactlyOnce is not supported*); Last>] Cosmos of ParseResults<DestinationParameters>
+        | [<CliPrefix(CliPrefix.None); Unique(*ExactlyOnce is not supported*); Last>] Es of ParseResults<EsSinkParameters>
+        | [<CliPrefix(CliPrefix.None); Unique(*ExactlyOnce is not supported*); Last>] Cosmos of ParseResults<CosmosSinkParameters>
         interface IArgParserTemplate with
             member a.Usage =
                 match a with
@@ -182,9 +179,14 @@ module CmdParser =
 #endif
                 | CategoryBlacklist _ ->    "category whitelist"
                 | CategoryWhitelist _ ->    "category blacklist"
-                | Cosmos _ ->               "CosmosDb destination parameters."
+                | Es _ ->                   "EventStore Sink parameters."
+                | Cosmos _ ->               "CosmosDb Sink parameters."
     and SourceArguments(a : ParseResults<SourceParameters>) =
-        member val Destination =        DestinationArguments(a.GetResult Cosmos)
+        member val Sink =
+            match a.TryGetSubCommand() with
+            | Some (Cosmos cosmos) -> Choice1Of2 (CosmosSinkArguments cosmos)
+            | Some (Es es) -> Choice2Of2 (EsSinkArguments es)
+            | _ -> raise (InvalidArguments "Must specify one of cosmos or es for Sink")
         member __.CategoryFilterFunction : string -> bool =
             match a.GetResults CategoryBlacklist, a.GetResults CategoryWhitelist with
             | [], [] ->     Log.Information("Not filtering by category"); fun _ -> true 
@@ -226,7 +228,7 @@ module CmdParser =
             GesConnector(__.User, __.Password, __.Timeout, __.Retries, log=log, heartbeatTimeout=__.Heartbeat, tags=tags)
                 .Establish("SyncTemplate", __.Discovery, connectionStrategy) |> Async.RunSynchronously
 #endif
-    and [<NoEquality; NoComparison>] DestinationParameters =
+    and [<NoEquality; NoComparison>] CosmosSinkParameters =
         | [<AltCommandLine("-s")>] Connection of string
         | [<AltCommandLine("-d")>] Database of string
         | [<AltCommandLine("-c")>] Collection of string
@@ -244,28 +246,65 @@ module CmdParser =
                 | Retries _ ->              "specify operation retries (default: 0)."
                 | RetriesWaitTime _ ->      "specify max wait-time for retry when being throttled by Cosmos in seconds (default: 5)"
                 | ConnectionMode _ ->       "override the connection mode (default: DirectTcp)."
-    and DestinationArguments(a : ParseResults<DestinationParameters>) =
+    and CosmosSinkArguments(a : ParseResults<CosmosSinkParameters>) =
         member __.Mode =                a.GetResult(ConnectionMode, Equinox.Cosmos.ConnectionMode.DirectTcp)
         member __.Discovery =           Discovery.FromConnectionString __.Connection
         member __.Connection =          match a.TryGetResult Connection  with Some x -> x | None -> envBackstop "Connection" "EQUINOX_COSMOS_CONNECTION"
         member __.Database =            match a.TryGetResult Database    with Some x -> x | None -> envBackstop "Database"   "EQUINOX_COSMOS_DATABASE"
         member __.Collection =          match a.TryGetResult Collection  with Some x -> x | None -> envBackstop "Collection" "EQUINOX_COSMOS_COLLECTION"
 
-        member __.Timeout =             a.GetResult(Timeout, 5.) |> TimeSpan.FromSeconds
-        member __.Retries =             a.GetResult(Retries, 0)
+        member __.Timeout =             a.GetResult(CosmosSinkParameters.Timeout, 5.) |> TimeSpan.FromSeconds
+        member __.Retries =             a.GetResult(CosmosSinkParameters.Retries, 0)
         member __.MaxRetryWaitTime =    a.GetResult(RetriesWaitTime, 5)
 
         /// Connect with the provided parameters and/or environment variables
         member x.Connect
             /// Connection/Client identifier for logging purposes
-            name : Async<CosmosConnection> =
+            appName connIndex : Async<CosmosConnection> =
             let (Discovery.UriAndKey (endpointUri,_masterKey)) as discovery = x.Discovery
             Log.Information("Destination CosmosDb {mode} {endpointUri} Database {database} Collection {collection}",
                 x.Mode, endpointUri, x.Database, x.Collection)
             Log.Information("Destination CosmosDb timeout {timeout}s; Throttling retries {retries}, max wait {maxRetryWaitTime}s",
                 (let t = x.Timeout in t.TotalSeconds), x.Retries, x.MaxRetryWaitTime)
             let c = CosmosConnector(x.Timeout, x.Retries, x.MaxRetryWaitTime, Log.Logger, mode=x.Mode)
-            c.Connect(name, discovery)
+            c.Connect(sprintf "App=%s Conn=%d" appName connIndex, discovery)
+    and [<NoEquality; NoComparison>] EsSinkParameters =
+        | [<AltCommandLine("-v")>] VerboseStore
+        | [<AltCommandLine("-o")>] Timeout of float
+        | [<AltCommandLine("-r")>] Retries of int
+        | [<AltCommandLine("-g")>] Host of string
+        | [<AltCommandLine("-x")>] Port of int
+        | [<AltCommandLine("-u")>] Username of string
+        | [<AltCommandLine("-p")>] Password of string
+        | [<AltCommandLine("-h")>] HeartbeatTimeout of float
+        interface IArgParserTemplate with
+            member a.Usage =
+                match a with
+                | VerboseStore ->           "Include low level Store logging."
+                | Timeout _ ->              "specify operation timeout in seconds (default: 20)."
+                | Retries _ ->              "specify operation retries (default: 3)."
+                | Host _ ->                 "specify a DNS query, using Gossip-driven discovery against all A records returned (defaults: envvar:EQUINOX_ES_HOST, localhost)."
+                | Port _ ->                 "specify a custom port (default: envvar:EQUINOX_ES_PORT, 30778)."
+                | Username _ ->             "specify a username (defaults: envvar:EQUINOX_ES_USERNAME, admin)."
+                | Password _ ->             "specify a password (defaults: envvar:EQUINOX_ES_PASSWORD, changeit)."
+                | HeartbeatTimeout _ ->     "specify heartbeat timeout in seconds (default: 1.5)."
+    and EsSinkArguments(a : ParseResults<EsSinkParameters>) =
+        member __.Host =                match a.TryGetResult Host       with Some x -> x | None -> envBackstop "Host"       "EQUINOX_ES_HOST"
+        member __.Port =                match a.TryGetResult Port       with Some x -> Some x | None -> Environment.GetEnvironmentVariable "EQUINOX_ES_PORT" |> Option.ofObj |> Option.map int
+        member __.Discovery =           match __.Port                   with Some p -> Discovery.GossipDnsCustomPort (__.Host, p) | None -> Discovery.GossipDns __.Host 
+        member __.User =                match a.TryGetResult Username   with Some x -> x | None -> envBackstop "Username"   "EQUINOX_ES_USERNAME"
+        member __.Password =            match a.TryGetResult Password   with Some x -> x | None -> envBackstop "Password"   "EQUINOX_ES_PASSWORD"
+        member __.Heartbeat =           a.GetResult(HeartbeatTimeout,1.5) |> TimeSpan.FromSeconds
+        member __.Timeout =             a.GetResult(Timeout,20.) |> TimeSpan.FromSeconds
+        member __.Retries =             a.GetResult(Retries,3)
+        member __.Connect(log: ILogger, storeLog: ILogger, connectionStrategy, appName, connIndex) =
+            let s (x : TimeSpan) = x.TotalSeconds
+            log.Information("EventStore {host} Connection {connId} heartbeat: {heartbeat}s Timeout: {timeout}s Retries {retries}",
+                __.Host, connIndex, s __.Heartbeat, s __.Timeout, __.Retries)
+            let log=if storeLog.IsEnabled Serilog.Events.LogEventLevel.Debug then Logger.SerilogVerbose storeLog else Logger.SerilogNormal storeLog
+            let tags=["M", Environment.MachineName; "I", Guid.NewGuid() |> string; "App", appName; "Conn", connIndex]
+            GesConnector(__.User, __.Password, __.Timeout, __.Retries, log=log, heartbeatTimeout=__.Heartbeat, tags=tags)
+                .Establish("SyncTemplate", __.Discovery, connectionStrategy)
 
     /// Parse the commandline; can throw exceptions in response to missing arguments and/or `-h`/`--help` args
     let parse argv : Arguments =
@@ -286,29 +325,86 @@ module Logging =
                         let cfpLevel = if verboseConsole then LogEventLevel.Debug else LogEventLevel.Warning
                         c.MinimumLevel.Override("Microsoft.Azure.Documents.ChangeFeedProcessor", cfpLevel)
             |> fun c -> let ingesterLevel = if verboseConsole then LogEventLevel.Debug else LogEventLevel.Information
-                        c.MinimumLevel.Override(typeof<Equinox.Projection.Scheduling.StreamStates>.FullName, ingesterLevel)
+                        c.MinimumLevel.Override(typeof<Propulsion.Streams.Scheduling.StreamStates<_>>.FullName, ingesterLevel)
             |> fun c -> if verbose then c.MinimumLevel.Debug() else c
             |> fun c -> let generalLevel = if verbose then LogEventLevel.Information else LogEventLevel.Warning
-                        c.MinimumLevel.Override(typeof<CosmosSink.Writer.Result>.FullName, generalLevel)
+                        c.MinimumLevel.Override(typeof<Propulsion.Cosmos.Internal.Writer.Result>.FullName, generalLevel)
+                         .MinimumLevel.Override(typeof<Propulsion.EventStore.Internal.Writer.Result>.FullName, generalLevel)
                          .MinimumLevel.Override(typeof<Checkpoint.CheckpointSeries>.FullName, LogEventLevel.Information)
             |> fun c -> let t = "[{Timestamp:HH:mm:ss} {Level:u3}] {partitionKeyRangeId} {Tranche} {Message:lj} {NewLine}{Exception}"
                         let configure (a : Configuration.LoggerSinkConfiguration) : unit =
                             a.Logger(fun l ->
-                                l.WriteTo.Sink(Equinox.Cosmos.Store.Log.InternalMetrics.RuCounters.RuCounterSink()) |> ignore) |> ignore
+                                l.WriteTo.Sink(Equinox.EventStore.Log.InternalMetrics.Stats.LogSink())
+                                 .WriteTo.Sink(Equinox.Cosmos.Store.Log.InternalMetrics.Stats.LogSink()) |> ignore) |> ignore
                             a.Logger(fun l ->
                                 let isEqx = Filters.Matching.FromSource<Core.CosmosContext>().Invoke
+                                let isWriterA = Filters.Matching.FromSource<Propulsion.EventStore.Internal.Writer.Result>().Invoke
+                                let isWriterB = Filters.Matching.FromSource<Propulsion.Cosmos.Internal.Writer.Result>().Invoke
                                 let isCp = Filters.Matching.FromSource<Checkpoint.CheckpointSeries>().Invoke
                                 let isWriter = Filters.Matching.FromSource<CosmosSink.Writer.Result>().Invoke
                                 let isCfp429a = Filters.Matching.FromSource("Microsoft.Azure.Documents.ChangeFeedProcessor.LeaseManagement.DocumentServiceLeaseUpdater").Invoke
                                 let isCfp429b = Filters.Matching.FromSource("Microsoft.Azure.Documents.ChangeFeedProcessor.PartitionManagement.LeaseRenewer").Invoke
                                 let isCfp429c = Filters.Matching.FromSource("Microsoft.Azure.Documents.ChangeFeedProcessor.PartitionManagement.PartitionLoadBalancer").Invoke
-                                (if verboseConsole then l else l.Filter.ByExcluding(fun x -> isEqx x || isCp x || isWriter x || isCfp429a x || isCfp429b x || isCfp429c x))
+                                let isCfp429d = Filters.Matching.FromSource("Microsoft.Azure.Documents.ChangeFeedProcessor.FeedProcessing.PartitionProcessor").Invoke
+                                let isCfp x = isCfp429a x || isCfp429b x || isCfp429c x || isCfp429d x
+                                (if verboseConsole then l else l.Filter.ByExcluding(fun x -> isEqx x || isWriterA x || isWriterB x || isCfp x))
                                     .WriteTo.Console(theme=Sinks.SystemConsole.Themes.AnsiConsoleTheme.Code, outputTemplate=t)
                                     |> ignore) |> ignore
                         c.WriteTo.Async(bufferSize=65536, blockWhenFull=true, configure=Action<_> configure)
             |> fun c -> match maybeSeqEndpoint with None -> c | Some endpoint -> c.WriteTo.Seq(endpoint)
             |> fun c -> c.CreateLogger()
-        Log.ForContext<Equinox.Projection.Scheduling.StreamStates>(), Log.ForContext<Core.CosmosContext>()
+        Log.ForContext<Propulsion.Streams.Scheduling.StreamStates<_>>(), Log.ForContext<Core.CosmosContext>()
+
+ //#if marveleqx
+[<RequireQualifiedAccess>]
+module EventV0Parser =
+    open Newtonsoft.Json
+
+    /// A single Domain Event as Written by internal Equinox versions
+    type [<NoEquality; NoComparison; JsonObject(ItemRequired=Required.Always)>]
+        EventV0 =
+        {   /// DocDb-mandated Partition Key, must be maintained within the document
+            s: string // "{streamName}"
+
+            /// Creation datetime (as opposed to system-defined _lastUpdated which is touched by triggers, replication etc.)
+            c: DateTimeOffset // ISO 8601
+
+            /// The Case (Event Type); used to drive deserialization
+            t: string // required
+
+            /// 'i' value for the Event
+            i: int64 // {index}
+
+            /// Event body, as UTF-8 encoded json ready to be injected into the Json being rendered for DocDb
+            [<JsonConverter(typeof<Equinox.Cosmos.Internal.Json.VerbatimUtf8JsonConverter>)>]
+            d: byte[] }
+
+    type Microsoft.Azure.Documents.Document with
+        member document.Cast<'T>() =
+            let tmp = new Microsoft.Azure.Documents.Document()
+            tmp.SetPropertyValue("content", document)
+            tmp.GetPropertyValue<'T>("content")
+
+    /// Maps fields in an Equinox V0 Event to the interface defined by the Propulsion.Streams library
+    let (|StandardCodecEvent|) (x: EventV0) = Propulsion.Streams.Internal.EventData.Create(x.t, x.d, timestamp = x.c)
+
+    /// We assume all Documents represent Events laid out as above
+    let parse (d : Microsoft.Azure.Documents.Document) : Propulsion.Streams.StreamEvent<_> =
+        let (StandardCodecEvent e) as x = d.Cast<EventV0>()
+        { stream = x.s; index = x.i; event = e } : _
+
+let transformV0 categorize catFilter (v0SchemaDocument: Microsoft.Azure.Documents.Document) : Propulsion.Streams.StreamEvent<_> seq = seq {
+    let parsed = EventV0Parser.parse v0SchemaDocument
+    let streamName = (*if parsed.Stream.Contains '-' then parsed.Stream else "Prefixed-"+*)parsed.stream
+    if catFilter (categorize streamName) then
+        yield parsed }
+//#else
+let transformOrFilter categorize catFilter (changeFeedDocument: Microsoft.Azure.Documents.Document) : Propulsion.Streams.StreamEvent<_> seq = seq {
+    for { stream = s; index = i; event = e } in DocumentParser.enumEvents changeFeedDocument do
+        // NB the `index` needs to be contiguous with existing events - IOW filtering needs to be at stream (and not event) level
+        if catFilter (categorize s) then
+            yield { stream = s; index = i; event = Propulsion.Streams.Internal.EventData.Create(e.EventType, e.Data, e.Meta, e.Timestamp) } }
+//#endif
 
 [<EntryPoint>]
 let main argv =